feat: enhance authentication and group management functionality

+++ backend/src/group/group.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
   NotFoundException,
@@ -12,7 +13,7 @@ export class GroupService {
   constructor(private readonly prisma: PrismaService) {}
 
   async validateUser(userId: string) {
-    const user = await this.prisma.group.findUnique({
+    const user = await this.prisma.user.findUnique({
       where: { id: userId },
     });
     if (!user) throw new NotFoundException('No se encontro el usuario');
@@ -36,6 +37,12 @@ export class GroupService {
 
   async create(createGroupDto: Omit<CreateGroupDto, 'userId'>, userId: string) {
     await this.validateUser(userId);
+    const existingGroup = await this.prisma.group.findFirst({
+      where: { name: createGroupDto.name, userId },
+    });
+    if (existingGroup) {
+      throw new BadRequestException('Ya existe un grupo con ese nombre');
+    }
     return this.prisma.group.create({
       data: { ...createGroupDto, userId },
     });
@@ -58,6 +65,14 @@ export class GroupService {
 
   async update(id: string, updateGroupDto: UpdateGroupDto, userId: string) {
     await this.validateGroup(id, userId);
+    if (updateGroupDto.name) {
+      const existingGroup = await this.prisma.group.findFirst({
+        where: { name: updateGroupDto.name, userId },
+      });
+      if (existingGroup && existingGroup.id !== id) {
+        throw new BadRequestException('Ya existe un grupo con ese nombre');
+      }
+    }
 
     return this.prisma.group.update({
       where: { id },
@@ -67,6 +82,13 @@ export class GroupService {
 
   async remove(id: string, userId: string) {
     await this.validateGroup(id, userId);
+    // Si el grupo tiene notas, no se puede eliminar
+    const notes = await this.prisma.note.findMany({
+      where: { groupId: id },
+    });
+    if (notes.length > 0) {
+      throw new BadRequestException('No se puede eliminar un grupo con notas');
+    }
 
     return this.prisma.group.delete({ where: { id } });
   }
